+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Bug fixes:
 Improvements:
  * icpp: automatically generate the standard module pcm file in user local system;
  * imod: add assets configuration to support icpp module extra payloads;
+ * imod: make an os independent module if there's no extra include paths;
 
 ## v0.1.1
 Bug fixes: